% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsModelSpec.R
\name{harmonic}
\alias{harmonic}
\title{Create a sine/cosine basis}
\usage{
harmonic(x, nfreq, period, intercept = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{nfreq}{number of sine/cosine pairs to include}

\item{period}{the period}

\item{intercept}{should basis matrix include a column of 1s?}
}
\description{
Create a matrix of sine and cosine basis vectors
}
