% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsModelSpec.R
\name{Lag}
\alias{Lag}
\title{Create Lagged Variables}
\usage{
Lag(v, k, group = NULL)
}
\arguments{
\item{v}{a numeric vector}

\item{k}{an integer vector giving lag numbers}

\item{group}{a factor or a list of factors defining groups of observations}
}
\description{
Create a matrix of lagged variables
}
\examples{
## Ten day "time series"
x <- rnorm(10)

## Lag 1 of `x'
Lag(x, 1)

## Lag 0, 1, and 2 of `x'
Lag(x, 0:2)

}
