% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsModelSpec.R
\name{runMean}
\alias{runMean}
\title{Compute Running Means}
\usage{
runMean(v, lags = 0, group = NULL, filter = NULL)
}
\arguments{
\item{v}{a numeric vector}

\item{lags}{an integer vector giving lag numbers}

\item{group}{a factor or a list of factors defining groups of observations}

\item{filter}{a vector specifying a linear filter}
}
\description{
Compute the running mean of a vector
}
\examples{
## Ten day "time series"
x <- rnorm(10)

## Running mean of lag 0, 1, and 2
runMean(x, 0:2)

}
